backend/framework/requirements.txt
aiohttp
aioredis
aioredis-cluster
aiosignal
anyio
aioredis
boto3
botocore
firebase-admin
httpcore
httpx
imgkit
Jinja2
pandas
pydantic
PyJWT
PyYAML
redis
requests
jsonbender
es-pandas
openpyxl
opensearch-py
gevent
fastapi
email-validator
elasticsearch[async]==7.13.1
uvicorn
redis
pydantic[dotenv]
pyrebase4